+++ tests/challenges/basic_abilities/goal_oriented_tasks.md
