+++ tests/challenges/challenge_decorator/challenge.py
