+++ tests/challenges/test_challenge_should_be_formatted_properly.py
