+++ tests/unit/test_agent_manager.py
