+++ tests/unit/test_ai_config.py
