+++ tests/unit/test_api_manager.py
